This is an initial commit of three matlab files
cycle_check.m serves to find if adding a new branch will create a cycle in our tree

cycle_check_start.m serves to find the optimal node in a potenitial branch to begin looking for cycles.

mutual_information_computation_10_7_16.m is the main script to create a graph of the estimated grid topology.